Round cost products so every architecture ranks candidates alike
The encoder picks the cheapest of several candidate encodings by comparing
float64 cost estimates, and candidates routinely score within an ulp of each
other. The Go spec lets a backend fuse a multiply into the add or subtract
that consumes it; arm64 does and amd64 does not, so cost loops of the form
`sum -= float64(count) * log2(count)` summed differently per architecture,
flipped the `cost < bestCost` comparison, and produced different bytes. Three
of the five images in testdata/transparent encoded to different ALPH chunks on
arm64 and amd64, differing by up to 0.35% of the file.

Rounding each product through an explicit conversion stops the fusion, which
is all that was needed: everything else in the cost math is plain IEEE-754 and
already reproduces exactly. arm64 now produces byte-for-byte what amd64
already produced, so no existing output changes on amd64.

All four sites the arm64 backend was fusing are covered, including the two
lookup tables built at init, whose entries could have landed on either side of
a rounding boundary.

Also correct the stale claim above elossyAlphaFilterCost that entropy ranking
is how libwebp picks a filter. The comment above
elossyAlphaFilterCandidates explains entropy ranking is a poor proxy for what
VP8L charges, which is why the higher effort levels trial-encode instead.

+++ enc_alpha.go
@@ -93,8 +93,8 @@ func elossyFilterAlphaPlane(alpha []byte, filter uint8, width, height int) []byt
 
 // elossyAlphaFilterCost estimates what a filtered plane costs to store, as the
 // zero-order entropy of its bytes in bits. Ranking the four filters this way
-// costs one pass each instead of four trial compressions, which is how libwebp
-// picks a filter too.
+// costs one pass each instead of four trial compressions, which is why the
+// cheap effort levels use it to narrow the field before trial encoding.
 func elossyAlphaFilterCost(filtered []byte) float64 {
 	var histogram [256]int
 	for _, v := range filtered {
@@ -107,7 +107,9 @@ func elossyAlphaFilterCost(filtered []byte) float64 {
 			continue
 		}
 		p := float64(count) / total
-		cost -= float64(count) * math.Log2(p)
+		// The conversion keeps the product out of a fused multiply-subtract.
+		// See enc_fma_test.go.
+		cost -= float64(float64(count) * math.Log2(p))
 	}
 	return cost
 }
